+++ README.md
@@ -1,22 +1,12 @@
-# Boomerang Worker Interfaces
+# Boomerang Worker Interface
 
-The following interfaces provide integrations to the Boomerang Workers, in particular for Boomerang Flow, by providing an easy to leverage CLI and also Common Core functions or methods.
+The following interface provide an opinionated CLI for the the Boomerang Workers, in particular for Boomerang Flow, by providing an easy to leverage CLI with core functionality
 
-### CLI
-
-The CLI is a main execution script that offers task developers the ease of just having to write a `.js` file to do a specific command.
-
-### Core
-
-The core module containers the following main methods
-
-| Method | Purpose                                                                                                             |
-| ------ | ------------------------------------------------------------------------------------------------------------------- |
-| common | Collection of common methods that can be used by task developers to enhance and speed the ease of their development |
-| log    | Collection of logging utilities using chalk to output nice values in the log for the user                           |
-| utils  | Collection of utility functions to help plugin authors retrieve, resolve, and set properties.                       |
-
-In addition the following files are used: config.js and index.js
+- Provides task developers the ease of just having to write a `.js` file to do a specific command.
+- Provides a set of utilities to help task developers speed up their development
+  - Common methods that can be used by task developers to enhance and speed the ease of their development
+  - Logging utilities using chalk to output nice values in the log for the user
+  - Collection of utility functions to help plugin authors retrieve, resolve, and set params
 
 ## Contributing
 
@@ -26,11 +16,7 @@ Do you want to contribute your own worker? Follow our [Contributing a Worker Gui
 
 ## Design
 
-The CLI has a main `cli.js` which imports all the `*.js` files under `./commands` folder. These are then mapped to the task / plugins command that are sent through as arguments on the `flow_task_template mongodb` collection. A command and sub command are required for all runs.
-
-### Project Structure
-
-Uses [pnpm workspaces](https://pnpm.io/workspaces) and [lerna](https://github.com/lerna/lerna) to manage the project as a monorepo.
+The CLI has a main `cli.js` which imports all the `*.js` files under `./commands` folder. These are then mapped to the task / plugins command that are sent through as arguments on the Task. A command and sub command are required for all runs.
 
 ### Handling Failure
 
@@ -42,41 +28,17 @@ The following command will use lerna to release new versions of the modules. It 
 
 `pnpm release`
 
-## Clean up
+## Local Development
+
+_TBA_
+
+## Troubleshooting
 
 ### Cleaning up jobs in Kubernetes
 
 When running against the non production cluster. You will need to clean up your runs using `kubectl delete job -l "app=bmrg-flow"`
 
-## References
-
-- [Node Initial Starter Tutorial](https://scotch.io/tutorials/build-an-interactive-command-line-application-with-nodejs)
-
-## Local Development
-
-Go to CLI package and install it globally for use outside of project. Link for use inside another project that has it as a dependency.
-
-```sh
-cd packages/cli
-yarn install -g
-yarn link
-```
-
-Open shell and run command to execute CLI
-
-```sh
-boomerang-worker-cli
-```
-
-If you want to test changes in another projectd that uses the CLI, in the root of the project run the following command:
-
-```sh
-yarn link "@boomerang-worker/cli"
-```
-
-## Troubleshooting
-
-**ESLint not working in VSCode**
+### ESLint not working in VSCode
 
 Try adding the following to your VSCode `settings.json` file
