+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@boomerang-io/boomerang-worker",
-  "version": "0.0.0",
+  "version": "1.0.0",
   "private": true,
   "repository": "git@github.com:boomerang-io/boomerang.worker.main.git",
   "author": "Tyson Lawrie <twlawrie@us.ibm.com>",
@@ -20,16 +20,6 @@
     "prepare": "husky install"
   },
   "devDependencies": {
-<<<<<<< Updated upstream
-    "@commitlint/cli": "^17.0.0",
-    "@commitlint/config-conventional": "^17.0.0",
-    "cross-env": "^7.0.2",
-    "eslint": "^8.18.0",
-    "eslint-plugin-node": "^11.1.0",
-    "husky": "^8.0.0",
-    "jest": "^28.1.1",
-    "lerna": "^5.1.4",
-=======
     "@commitlint/cli": "^19.2.0",
     "@commitlint/config-conventional": "^19.2.2",
     "cross-env": "^7.0.3",
@@ -38,7 +28,6 @@
     "husky": "^9.1.4",
     "jest": "^29.7.0",
     "lerna": "^8.1.7",
->>>>>>> Stashed changes
     "lerna-changelog": "^2.2.0",
     "lint-staged": "^15.2.8",
     "prettier": "^3.3.3"
