+++ packages/cli/package.json
@@ -1,19 +1,23 @@
 {
   "name": "@boomerang-io/worker-cli",
-  "version": "0.9.4",
+  "version": "1.0.0",
   "description": "Boomerang Worker CLI",
+  "homepage": "https://useboomerang.io",
   "bin": {
     "boomerang-worker-cli": "./bin/cli.js"
   },
-  "repository": "git@github.com:boomerang-io/boomerang.worker.main.git",
-  "author": "Tyson Lawrie <twlawrie@us.ibm.com>",
+  "type": "module",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/boomerang-io/worker.interfaces",
+    "directory": "packages/cli"
+  },
   "license": "Apache-2.0",
-  "type": "commonjs",
   "publishConfig": {
     "access": "public"
   },
   "engines": {
-    "node": ">=10"
+    "node": ">=18.0.0"
   },
   "scripts": {
     "dev:debug": "cross-env NODE_ENV=local DEBUG=true node ./bin/cli.js",
@@ -23,12 +27,10 @@
     "test:watch": "jest --watch"
   },
   "dependencies": {
-    "@boomerang-io/worker-core": "^1.3.11",
-    "commander": "^6.2.1",
+    "commander": "^12.1.0",
     "cross-spawn": "^7.0.3",
-    "dotenv": "^8.6.0",
-    "inquirer": "^7.3.3",
-    "node-datetime": "^2.1.2",
+    "dotenv": "^16.4.5",
+    "inquirer": "^10.1.7",
     "require-all": "^3.0.0",
     "shelljs": "^0.8.5"
   },
