+++ packages/cli/src/index.js
@@ -1,0 +1,2 @@
+import cli from './cli.js';
+export default cli;
