--- packages/core/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "@boomerang-io/worker-core",
-  "version": "1.3.11",
-  "description": "Boomerang Worker Core",
-  "main": "src/index.js",
-  "repository": "git@github.com:boomerang-io/boomerang.worker.main.git",
-  "author": "Tyson Lawrie <twlawrie@us.ibm.com>",
-  "license": "Apache-2.0",
-  "type": "commonjs",
-  "publishConfig": {
-    "access": "public"
-  },
-  "engines": {
-    "node": ">=10"
-  },
-  "scripts": {
-    "test": "echo 'test'"
-  },
-  "dependencies": {
-<<<<<<< Updated upstream
-    "app-root-path": "^3.0.0",
-    "chalk": "^4.1.2",
-    "node-datetime": "^2.1.2",
-    "node-fetch": "^2.6.0",
-=======
-    "app-root-path": "^3.1.0",
-    "chalk": "^5.0.1",
-    "date-fns": "^3.6.0",
->>>>>>> Stashed changes
-    "path": "^0.12.7",
-    "properties": "^1.2.1"
-  },
-  "devDependencies": {
-    "cross-env": "^7.0.2"
-  },
-  "gitHead": "878e09f06d7df2ea3d3c054e539ee405f30934f8"
-}
