--- packages/core/src/index.js
@@ -1,11 +1,0 @@
-const CICDError = require("./error");
-const log = require("./log");
-const utils = require("./utils");
-const common = require("./common");
-
-module.exports = {
-  CICDError,
-  common,
-  log,
-  utils,
-};
