--- pnpm-workspace.yaml
@@ -1,2 +1,0 @@
-packages:
-  - 'packages/*'
